+++ esphome/components/esp32_pm/pm.cpp
@@ -15,7 +15,7 @@
 #include <sstream>
 
 namespace esphome {
-namespace pm {
+namespace esp32_pm {
 
 static const char *const TAG = "PM";
 
@@ -34,9 +34,9 @@ void dump_locks() {
   }
 }
 
-void PM::setup() {
+void ESP32PowerManagement::setup() {
 #ifdef CONFIG_PM_ENABLE
-  ESP_LOGI(TAG, "PM Support Enabled");
+  ESP_LOGI(TAG, "ESP32_PM Support Enabled");
   esp_pm_config_esp32_t pm_config;
   pm_config.max_freq_mhz = max_freq_;
   pm_config.min_freq_mhz = min_freq_;
@@ -72,7 +72,7 @@ void PM::setup() {
   global_pm = this;
 }
 
-void PM::loop() {
+void ESP32PowerManagement::loop() {
   // Disable the startup lock once we have looped once
   if (this->setup_done_) {
     esp_pm_lock_release(this->startup_lock_);
@@ -80,7 +80,7 @@ void PM::loop() {
   }
 }
 
-void PM::dump_config() {
+void ESP32PowerManagement::dump_config() {
 #ifdef CONFIG_PM_ENABLE
   ESP_LOGCONFIG(TAG, "PM Support Enabled");
   ESP_LOGCONFIG(TAG, "Setting Minimum frequency to %dMHz, Maximum to %dMHz", min_freq_, max_freq_);
@@ -99,14 +99,14 @@ void PM::dump_config() {
 #endif  // CONFIG_PM_ENABLE
 }
 
-void PM::set_freq(uint16_t min_freq_mhz, uint16_t max_freq_mhz) {
+void ESP32PowerManagement::set_freq(uint16_t min_freq_mhz, uint16_t max_freq_mhz) {
   this->min_freq_ = min_freq_mhz;
   this->max_freq_ = max_freq_mhz;
 }
 
-void PM::set_tickless(bool tickless) { this->tickless_ = tickless; }
+void ESP32PowerManagement::set_tickless(bool tickless) { this->tickless_ = tickless; }
 
-std::unique_ptr<pm::PMLock> PM::get_lock() { return make_unique<PMLock>(this->pm_lock_); }
+std::unique_ptr<esp32_pm::PMLock> ESP32PowerManagement::get_lock() { return make_unique<PMLock>(this->pm_lock_); }
 
 PMLock::PMLock(std::shared_ptr<esp_pm_lock_handle_t> pm_lock) {
 #ifdef CONFIG_PM_ENABLE
@@ -126,8 +126,8 @@ PMLock::~PMLock() {
 #endif
 }
 
-PM *global_pm = nullptr;  // NOLINT(cppcoreguidelines-avoid-non-const-global-variables)
+ESP32PowerManagement *global_pm = nullptr;  // NOLINT(cppcoreguidelines-avoid-non-const-global-variables)
 
-}  // namespace pm
+}  // namespace esp32_pm
 }  // namespace esphome
 #endif
