+++ esphome/components/esp32_pm/pm.h
@@ -8,7 +8,7 @@
 #include "esp_pm.h"
 
 namespace esphome {
-namespace pm {
+namespace esp32_pm {
 
 class PMLock {
  public:
@@ -21,16 +21,16 @@ class PMLock {
 #endif
 };
 
-class PM : public Component {
+class ESP32PowerManagement : public Component {
  public:
   void setup() override;
   void set_freq(uint16_t min_freq_mhz, uint16_t max_freq_mhz);
-  float get_setup_priority() const { return setup_priority::BUS; }
+  float get_setup_priority() const override { return setup_priority::BUS; }
   void set_tickless(bool tickless);
-  void loop();
+  void loop() override;
   void dump_config() override;
 
-  std::unique_ptr<pm::PMLock> get_lock();
+  std::unique_ptr<esp32_pm::PMLock> get_lock();
 
  private:
   uint16_t min_freq_ = 40;
@@ -43,8 +43,8 @@ class PM : public Component {
 #endif
 };
 
-extern PM *global_pm;  // NOLINT(cppcoreguidelines-avoid-non-const-global-variables)
+extern ESP32PowerManagement *global_pm;  // NOLINT(cppcoreguidelines-avoid-non-const-global-variables)
 
-}  // namespace pm
+}  // namespace esp32_pm
 }  // namespace esphome
 #endif
